requirements.txt
asgiref==3.2.10
Django==3.1.2
django-environ==0.4.5
django-wkhtmltopdf==3.3.0
djangorestframework==3.12.2
flake8==3.8.4
gunicorn==20.0.4
importlib-metadata==2.0.0
isort==5.6.4
mccabe==0.6.1
Pillow==8.0.1
pkg-resources==0.0.0
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
zipp==3.4.0
